+++ YOGO/src/components/atoms/IconAbsolute/style.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components/native';
 export const Container = styled.View`
   width: 100%;
   position: absolute;
-  top: 0;
-  left: -60;
+  top: 15%;
+  left: -16%;
   opacity: 0.4;
 `;
