+++ YOGO/src/components/organisms/ResultSheet/style.tsx
@@ -1,0 +1,18 @@
+import styled from 'styled-components/native';
+
+export const Inner = styled.View`
+  width: 100%;
+  flex: 1;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+`;
+
+
+export const ResultBox = styled.View`
+  height: 100%;
+  width: 100%;
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px 0
+`;
