+++ YOGO/src/components/organisms/SearchSheet/style.tsx
@@ -1,0 +1,23 @@
+import styled from 'styled-components/native';
+
+export const ScrollView = styled.ScrollView`
+  width: 100%;
+  height: 100%;
+  margin-bottom: 50px;
+`;
+
+export const Inner = styled.View`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+`
+
+export const SearchBox = styled.View`
+  height: 100%;
+  width: 100%;
+  justify-content: flex-start;
+  align-items: center;
+`;
