+++ YOGO/src/components/templates/BottomSheet/index.tsx
@@ -6,53 +6,27 @@ import {
   Dimensions,
   PanResponder,
 } from 'react-native';
-import { DateTimePickerEvent } from '@react-native-community/datetimepicker';
 import {
-  SearchTarget,
-  SelectTargetCityBtn,
-  SelectTargetDate,
-  HeaderCenter,
-  BottomSheetBtn,
+  ResultSheet,
+  SearchSheet
 } from 'components';
 import { IconBottomSheetBar } from 'assets';
-import { DUMMY_DATA_CITY } from 'utils';
 import * as S from './style';
 interface ISearchBSProps {
   modalVisible: boolean;
   setModalVisible: (visible: boolean) => void;
-  onPressResult: () => void;
 }
 
-export const SearchBottomSheet = ({
+export const BottomSheet = ({
   modalVisible,
   setModalVisible,
-  onPressResult,
 }: ISearchBSProps) => {
-  const [date, setDate] = useState(new Date());
-  const [text, setText] = useState('');
 
-  const [selectedSearchTargetCity, setSelectedSearchTargetCity] =
-    useState<boolean>(false);
+    const [result, setResult] =
+  useState<boolean>(false);
 
-  const targetList = DUMMY_DATA_CITY.filter(item =>
-    item.city.toUpperCase().includes(text.toUpperCase()),
-  );
   const screenHeight = Dimensions.get('screen').height;
   const panY = useRef(new Animated.Value(screenHeight)).current;
-
-  const onChangeText = (text: string) => {
-    setText(text);
-  };
-
-  const onPressSearchTargetCity = () => {
-    setSelectedSearchTargetCity(true);
-    setText('');
-  };
-
-  const onSubmitText = (city: string) => {
-    setSelectedSearchTargetCity(false);
-    setText(city);
-  };
 
   const translateY = panY.interpolate({
     inputRange: [-1, 0, 1],
@@ -88,11 +62,6 @@ export const SearchBottomSheet = ({
     }),
   ).current;
 
-  const onChangeDate = (event: DateTimePickerEvent, selectedDate: Date) => {
-    const currentDate = selectedDate;
-    setDate(currentDate);
-  };
-
   useEffect(() => {
     if (modalVisible) {
       resetBottomSheet.start();
@@ -102,13 +71,15 @@ export const SearchBottomSheet = ({
   const closeModal = () => {
     closeBottomSheet.start(() => {
       setModalVisible(false);
-      setSelectedSearchTargetCity(false);
+      setResult(false);
     });
   };
 
-  const onPressBottomSheetBtn = () => {
-    closeModal();
-    onPressResult();
+  const onPressBottomSheetFindBtn = () => {
+    setResult(true);
+  }
+  const onPressBottomSheetMakeBtn = () => {
+    setResult(true);
   }
 
   return (
@@ -125,35 +96,13 @@ export const SearchBottomSheet = ({
 
         <S.Container
           height={screenHeight}
+          isResult={result}
           style={{ transform: [{ translateY: translateY }] }}
           {...panResponders.panHandlers}
         >
           <IconBottomSheetBar />
-          <S.SearchBox>
-            {!selectedSearchTargetCity ? (
-              <S.ScrollView showsVerticalScrollIndicator={false}>
-                <S.Inner>
-                  <HeaderCenter text={`Search Time Zone`} size={18} />
-
-                  <SelectTargetCityBtn
-                    onPress={() => onPressSearchTargetCity()}
-                    text={text.trim() === '' ? '국가, 도시' : text}
-                  />
-                  <SelectTargetDate onChangeDate={onChangeDate} date={date} />
-                  <BottomSheetBtn
-                    text={'FIND'}
-                    onPress={() => onPressBottomSheetBtn()}/>
-                </S.Inner>
-              </S.ScrollView>
-            ) : (
-              <SearchTarget
-                targetList={targetList}
-                text={text}
-                onChangeText={onChangeText}
-                onSubmitText={onSubmitText}
-              />
-            )}
-          </S.SearchBox>
+          {!result && (<SearchSheet onPress={onPressBottomSheetFindBtn}/>)}
+          {result && (<ResultSheet onPress={onPressBottomSheetMakeBtn}/>)}
         </S.Container>
       </S.Overlay>
     </Modal>
