+++ YOGO/src/components/templates/BottomSheet/style.tsx
@@ -1,0 +1,22 @@
+import styled from 'styled-components/native';
+import { Animated } from 'react-native';
+
+export const Container = styled(Animated.View)<{height: number, isResult: boolean}>`
+  height: ${({ height }) => height * 0.95}px;
+  justify-content: flex-start;
+  align-items: center;
+  background-color: ${({ isResult, theme }) => isResult ? theme.colors.blue : theme.colors.white};
+  border-top-left-radius: 10px;
+  border-top-right-radius: 10px;
+  padding-top: 10px;
+`;
+
+export const Background = styled.View`
+  flex: 1;
+`;
+
+export const Overlay = styled.View`
+  flex: 1;
+  justify-content: flex-end;
+  background-color: rgba(0, 0, 0, 0.4);
+`;
