+++ YOGO/src/components/templates/TimeZone/index.tsx
@@ -4,26 +4,18 @@ import {
   FloatingButton,
   HeaderRightButton,
   TimeZoneList,
-  SearchBottomSheet,
-  ResultBottomSheet,
+  BottomSheet,
 } from 'components';
 import { IconSearch } from 'assets';
 import * as S from './style';
 
 export function TimeZone() {
   const [modalVisible, setModalVisible] = useState<boolean>(false);
-  const [modalVisibleResult, setModalVisibleResult] = useState<boolean>(false);
- 
   const navigation = useNavigation();
 
   const pressBottomSheet = () => {
     setModalVisible(true);
   };
-
-  const pressBottomSheetResult = () => {
-    setTimeout(() => {
-      setModalVisibleResult(true)},1000)
-  }
 
   useEffect(() => {
     navigation.setOptions({
@@ -38,12 +30,10 @@ export function TimeZone() {
     <>
       <S.Container>
         <TimeZoneList />
-        <SearchBottomSheet
+        <BottomSheet
           modalVisible={modalVisible}
           setModalVisible={setModalVisible}
-          onPressResult = {pressBottomSheetResult}
         />
-        <ResultBottomSheet modalvisibleResult={modalVisibleResult} setmodalvisibleResult={setModalVisibleResult} />
       </S.Container>
       <FloatingButton onPress={() => pressBottomSheet()}>
         <IconSearch color="white" />
