+++ YOGO/src/components/templates/TimeZone/style.tsx
@@ -3,10 +3,3 @@ import styled from 'styled-components/native';
 export const Container = styled.View`
   background: #eee;
 `;
-
-export const Title = styled.Text`
-  color: blue;
-`;
-
-export const openBottomSheet = styled.Button`
-`
